% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bloodcontrol}
\alias{bloodcontrol}
\title{Blood samples from different individuals.}
\format{A data frame with 13 rows and 12 haematological variables:
\describe{
  \item{HCT}{haematocrit [\%]}
  \item{HGB}{haemoglobin [g/dL]}
  \item{IRF}{immature reticulocyte fraction [\%]}
  \item{MCH}{mean corpuscular haemoglobin [pg]}
  \item{MCHC}{mean corpuscular haemoglobin concentration [g/dL]}
  \item{MCV}{mean corpuscular volume [fL]}
  \item{RBC}{red blood cell count [10^6/uL]}
  \item{RDW.SD}{red blood cell distribution width [fL]}
  \item{RETC}{reticulocyte count [10^6/uL]}
  \item{RETP}{reticulocyte percentage [\%]}
  \item{OFFscore}{OFF-score}
  \item{ABPS}{Abnormal Blood Profile Score}
}

These samples are assumed to represent normal population.}
\source{
Swiss Laboratory for Doping Analyses (LAD), with some calculations
performed by the World Anti-Doping Agency (WADA).
}
\usage{
bloodcontrol
}
\description{
A dataset containing the result of the analysis of 13 blood samples
from different individuals.
}
\section{Note}{
 One of the rows actually belongs to one of the
    authors of this package, who promises that he was not doped.
}

\keyword{datasets}
